improve hyperparam parsing for model saving
